+++ src/helpers/FileHelper.php
@@ -4,7 +4,7 @@ namespace escape\escapedam\helpers;
 
 use Craft;
 
-class FileHelper
+final class FileHelper
 {
     /**
      * @param $fileUrl
@@ -12,56 +12,53 @@ class FileHelper
      * @return bool
      * @throws \Exception
      */
-    public static function downloadFile($fileUrl, $filePath)
+    public static function downloadFile($fileUrl, $filePath): bool
     {
 
-        $httpStatus = null;
-        $errorMessage = null;
-
-        if (\function_exists('curl_init')) {
-
-            $ch = \curl_init($fileUrl);
-            $fp = \fopen($filePath, "wb");
-
-            $options = [
-                CURLOPT_FILE => $fp,
-                CURLOPT_HEADER => 0,
-                CURLOPT_FOLLOWLOCATION => 1,
-                CURLOPT_TIMEOUT => 60,
-            ];
-
-            \curl_setopt_array($ch, $options);
-            \curl_exec($ch);
-
-            if (\curl_errno($ch) !== 0) {
-                $errorMessage = \curl_error($ch);
-            }
-
-            $httpStatus = (int) \curl_getinfo($ch, CURLINFO_HTTP_CODE);
-
-            \curl_close($ch);
-            \fclose($fp);
-
-        } elseif (\ini_get('allow_url_fopen')) {
-
-            \file_put_contents($filePath, \file_get_contents($fileUrl));
-            $httpStatus = $http_response_header[0] ?? null;
-        } else {
-
-            throw new \Exception('Looks like allow_url_fopen is off and cURL is not enabled. To download external files, one of these methods has to be enabled.');
+        if (empty($fileUrl)) {
+            throw new \Exception("File url cannot be empty");
         }
 
-        if ($errorMessage) {
+        $errorMessage = null;
+
+        if (!function_exists('curl_init')) {
+            throw new \Exception('Curl not installed');
+        }
+
+        $ch = curl_init($fileUrl);
+        $fp = fopen($filePath, "wb");
+
+        $options = [
+            CURLOPT_FILE => $fp,
+            CURLOPT_HEADER => 0,
+            CURLOPT_FOLLOWLOCATION => 1,
+            CURLOPT_TIMEOUT => 60,
+            CURLOPT_REFERER => Craft::$app->getSites()->getPrimarySite()->getBaseUrl(),
+        ];
+
+        curl_setopt_array($ch, $options);
+        curl_exec($ch);
+
+        if (curl_errno($ch) !== 0) {
+            $errorMessage = curl_error($ch);
+        }
+
+        $httpStatus = (int) curl_getinfo($ch, CURLINFO_HTTP_CODE);
+
+        curl_close($ch);
+        fclose($fp);
+
+        if (!empty($errorMessage)) {
             throw new \Exception(Craft::t('site', 'An error “{errorMessage}” occurred while attempting to download “{fileUrl}”', [
                 'fileUrl' => $fileUrl,
-                'errorMessage' => $errorMessage
+                'errorMessage' => $errorMessage,
             ]));
         }
 
         if ($httpStatus !== 200) {
             throw new \Exception(Craft::t('site', 'HTTP status “{httpStatus}” encountered while attempting to download “{fileUrl}”', [
                 'fileUrl' => $fileUrl,
-                'httpStatus' => $httpStatus
+                'httpStatus' => $httpStatus,
             ]));
         }