+++ src/assetbundles/fields/escapedam/resources/EscapeDamInputSelect.js
@@ -14,49 +14,125 @@ Craft.EscapeDam.DamSelectInput = Craft.AssetSelectInput.extend({
     damModal: null,
     disabledFileIds: null,
 
+    init: function() {
+        this.super.init.apply(this, arguments);
+
+        const $addNativeAssetsBtn = this.getAddNativeAssetsBtn();
+        if ($addNativeAssetsBtn.length) {
+            this.addListener($addNativeAssetsBtn, 'activate', 'showModal');
+        }
+    },
+
     getAddElementsBtn: function() {
-        return this.$container.find('.btn[data-input]');
+        return this.$container.find('.btn[data-input="dam"]:first');
+    },
+
+    getAddNativeAssetsBtn: function() {
+        return this.$container.find('.btn[data-input="assets"]:first');
     },
 
     disableAddElementsBtn: function() {
-        var $btn = this.$container.find('.btn[data-input]');
-        if ($btn) {
-            $btn.addClass('hidden');
-        }
+        this.$container.find('.btn[data-input]').each(function () {
+            $(this).addClass('hidden');
+        });
+
+        this.updateButtonContainer();
     },
 
     enableAddElementsBtn: function() {
-        var $btn = this.$container.find('.btn[data-input]');
-        if ($btn) {
-            $btn.removeClass('hidden');
+        if (this.settings.allowAdd) {
+            this.$container.find('.btn[data-input]').each(function () {
+                $(this).removeClass('hidden');
+            });
+        }
+        this.updateButtonContainer();
+    },
+
+    focusNextLogicalElement: function () {
+        if (!this.canAddMoreElements()) {
+            // If can add more elements, focus on add button
+            if (
+                this.$addElementBtn.length &&
+                !this.$addElementBtn.hasClass('hidden')
+            ) {
+                this.$addElementBtn.focus();
+            }
+        } else {
+            // If can't add more elements, focus on the final remove
+            this.focusLastRemoveBtn();
         }
     },
 
-    showModal: function (e) {
-        if (!this.canAddMoreElements()) {
-            return;
+    defineElementActions: function($element) {
+        // Remove the "Replace" action from assets in Escape DAM fields, because it doesn't work.
+        const actions = this.super.defineElementActions.apply(this, arguments)
+            .filter(action => action.label !== Craft.t('app', 'Replace'));
+
+        if (this.settings.showActionMenu) {
+            // ...and replace it with our own actions
+            actions.push({
+                icon: async () => await Craft.ui.icon('arrows-rotate'),
+                label: Craft.t('escapedam', 'Replace from DAM'),
+                callback: () => {
+                    this._$replaceElement = $element;
+                    this.showDamModal();
+                }
+            });
+
+            if (this.settings.enableAssetsInput) {
+                actions.push({
+                    icon: async () => await Craft.ui.icon('arrows-rotate'),
+                    label: Craft.t('escapedam', 'Replace from assets'),
+                    callback: () => {
+                        this._$replaceElement = $element;
+                        this.showNativeModal();
+                    }
+                });
+            }
         }
+
+        return actions;
+    },
+
+    showModal: function (e) {
         var $target = $(e.target);
         if ($target.data('input') === 'assets') {
             // Show default Assets modal
-            this.super.showModal.apply(this, arguments);
+            this.showNativeModal();
         } else {
             // Show super-awesome DAM modal
-            var _this = this;
-            if (!this.damModal) {
-                this.damModal = this.createDamModal({
-                    storageKey: window.location.pathname + '.' + this.settings.fieldId,
-                    onSelect: $.proxy(this.onDamModalSelect, this),
-                    onHide: function() {
-                        _this.$addElementBtn.focus();
-                        console.log('modal closed');
-                    },
-                    disabledFileIds: null, // TODO
-                    allowedExtensions: this.settings.allowedExtensions
-                });
-            } else {
-                this.damModal.show();
-            }
+            this.showDamModal();
+        }
+    },
+
+    showNativeModal() {
+        // Make sure we haven't reached the limit
+        if (!this._$replaceElement && !this.canAddMoreElements() || !this.settings.enableAssetsInput) {
+            return;
+        }
+        this.super.showModal.apply(this, arguments);
+    },
+
+    showDamModal() {
+        // Make sure we haven't reached the limit
+        if (!this._$replaceElement && !this.canAddMoreElements()) {
+            return;
+        }
+        if (!this.damModal) {
+            this.damModal = this.createDamModal({
+                storageKey: `${window.location.pathname}.${this.settings.fieldId}`,
+                onSelect: $.proxy(this.onDamModalSelect, this),
+                onHide: () => {
+                    this.$addElementBtn.focus({ preventScroll: true });
+                    if (!this.fileIdsToImport) {
+                        this._$replaceElement = null;
+                    }
+                },
+                disabledFileIds: null, // TODO
+                allowedExtensions: this.settings.allowedExtensions
+            });
+        } else {
+            this.damModal.show();
         }
     },
 
@@ -65,6 +141,9 @@ Craft.EscapeDam.DamSelectInput = Craft.AssetSelectInput.extend({
         if (this.settings.limit) {
             // Cut off any excess elements
             var slotsLeft = this.settings.limit - this.$elements.length;
+            if (this._$replaceElement) {
+                slotsLeft += 1;
+            }
             if (fileIds.length > slotsLeft) {
                 fileIds = fileIds.slice(0, slotsLeft);
             }
@@ -144,11 +223,12 @@ Craft.EscapeDam.DamSelectInput = Craft.AssetSelectInput.extend({
                                             ? 'chip'
                                             : 'card',
                                         size: this.settings.viewMode === 'large' ? 'large' : 'small',
-                                    },
-                                ],
-                            },
-                        ],
-                    },
+                                        showActionMenu: true
+                                    }
+                                ]
+                            }
+                        ]
+                    }
                 })
                     .then(async ({data}) => {
                         const elementInfo = Craft.getElementInfo(
@@ -180,13 +260,19 @@ Craft.EscapeDam.DamSelectInput = Craft.AssetSelectInput.extend({
     },
 
     _onImportComplete: function () {
+        if (this._$replaceElement) {
+            this.removeElement(this._$replaceElement);
+        }
+
+        this._$replaceElement = null;
+
         // Last file?
         if (this.fileIdsToImport.length) {
             this._importFile(this.fileIdsToImport.shift());
         } else {
             this.progressBar.hideProgressBar();
             this.$container.removeClass('uploading');
-            this.$addElementBtn.focus();
+            this.$addElementBtn.focus({ preventScroll: true });
         }
         Craft.cp.runQueue();
     },