+++ src/assetbundles/fields/escapedam/resources/EscapeDamInputSelect.js
@@ -219,10 +219,12 @@ Craft.EscapeDam.DamSelectInput = Craft.AssetSelectInput.extend({
                                 instances: [
                                     {
                                         context: 'field',
-                                        ui: ['list', 'large'].includes(this.settings.viewMode)
+                                        ui: ['list', 'list-inline', 'large', 'thumbs'].includes(this.settings.viewMode)
                                             ? 'chip'
                                             : 'card',
-                                        size: this.settings.viewMode === 'large' ? 'large' : 'small',
+                                        size: ['large', 'thumbs'].includes(this.settings.viewMode)
+                                          ? 'large'
+                                          : 'small',
                                         showActionMenu: true
                                     }
                                 ]